Update solution - CodeHub

+++ 8-kyu/keep-hydrated-1/keep-hydrated-1.java
@@ -1,6 +1,6 @@
  
 public class KeepHydrated  {
-  public static double liters(double time)  {
+  public static int liters(double time)  {
     return Math.floor(time * 0.5);
   }
 }
